Move MadeForYou and homePanel to separate files in components
